+++ database/migrations/2025_10_24_093236_make_deployable_url_nullable_in_deployments_table.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class() extends Migration {
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::table('gitlab_deployments', static function (Blueprint $table): void {
+            $table->string('deployable_url')
+                ->default('')
+                ->nullable()
+                ->change();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::table('gitlab_deployments', static function (Blueprint $table): void {
+            $table->string('deployable_url')
+                ->default('')
+                ->change();
+        });
+    }
+};