+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add indexes to the fields in the tables that are used in the repository to find the records if they exist.
 - Add Laravel 12 support.
 
+### Fixed
+- Make the deployable_url nullable on the deployments table because it can be null
+
 ## 0.6.0 - 2024-11-21
 
 ### Added